Added Delete function

+++ CodingSession.cs
@@ -7,6 +7,6 @@ namespace Code_Tracker
         public int Id { get; set; }
         public required string StartTime { get; set; }
         public required string EndTime { get; set; }
-        public double Duration { get; set; }
+        public float Duration { get; set; }
     }
 }
